Switch from Floats to uint64s for price quotes
The polling layer used to return float64s. This was changed, so now
all new datasources must be converted to uint64. This puts
the work at each datasource, but keeps the polling layer
consistent with the internals of pegnet. We should
continue to drop floats wherever we can. We should also verify
the json unmarshal of floats is accurate, or create a custom
unmarshal that reads the strings as fixed point